added some formation
