+++ src/components/tabbedpane/datatablesourse.js
@@ -1,15 +1,7 @@
 import { Link } from "react-router-dom";
 import React from "react";
 import EditableHeaderComponent from "./EditableHeaderComponent";
-import secureLocalStorage from "react-secure-storage";
-
-export const getIsStaff = () => {
-    if (secureLocalStorage.getItem("user_details")?.is_staff) {
-        return secureLocalStorage.getItem("user_details")?.is_staff;
-    } else {
-        return false;
-    }
-};
+import { getIsStaff } from "../../utils/common";
 
 export const defaultColDef = {
     tooltipValueGetter: (params) => {
