+++ src/components/tabbedpane/vlan/vlanTable.jsx
@@ -10,9 +10,10 @@ import VlanForm from "./VlanForm";
 import VlanMemberForm from "./vlanMemberForm";
 import VlanSagIpForm from "./vlanSagIpForm";
 import interceptor from "../../../utils/interceptor";
-import { getIsStaff } from "../datatablesourse";
+import { getIsStaff } from "../../../utils/common";
 import useStoreConfig from "../../../utils/configStore";
 import useStoreLogs from "../../../utils/store";
+import { isValidIPv4WithCIDR } from "../../../utils/common";
 
 // Function to get vlan names
 export const getVlanDataCommon = (selectedDeviceIp) => {
@@ -230,25 +231,6 @@ const VlanTable = (props) => {
         const selectedNodes = gridRef.current.api.getSelectedNodes();
         const selectedData = selectedNodes.map((node) => node.data);
         setSelectedRows(selectedData);
-    };
-
-    const isValidIPv4WithCIDR = (ipWithCidr) => {
-        if (ipWithCidr) {
-            const ipv4Regex =
-                /^(25[0-5]|2[0-4][0-9]|1[0-9][0-9]|[1-9]?[0-9])\.(25[0-5]|2[0-4][0-9]|1[0-9][0-9]|[1-9]?[0-9])\.(25[0-5]|2[0-4][0-9]|1[0-9][0-9]|[1-9]?[0-9])\.(25[0-5]|2[0-4][0-9]|1[0-9][0-9]|[1-9]?[0-9])$/;
-            const cidrRegex = /^([0-9]|[12][0-9]|3[0-2])$/;
-
-            const [ip, cidr] = ipWithCidr.split("/");
-
-            if (ipv4Regex.test(ip)) {
-                if (cidr === undefined || cidrRegex.test(cidr)) {
-                    return true;
-                }
-            }
-            return false;
-        } else {
-            return false;
-        }
     };
 
     const handleCellValueChanged = useCallback((params) => {
